+++ tergite/utils/job_file.py
@@ -10,11 +10,10 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-"""Functionality related to the job logfile"""
+"""Utility functions related to the job file"""
 import json
-from itertools import groupby
 from pathlib import Path
-from typing import Any, Dict, List, Tuple, Union
+from typing import Any, Dict
 
 import h5py
 
@@ -77,19 +76,3 @@ def _np_to_py(value: Any) -> Any:
         return value.item()
     except AttributeError:
         return value
-
-
-def iqx_rle(seq: List[Any]) -> List[Union[Tuple[Any], Tuple[Any, int]]]:
-    """Run-length encodes a sequence.
-
-    Constant subsequences are stored as a single term and count, rather than as the original subsequence.
-    Counts equal to 1 are omitted.
-
-    Args:
-        seq: the sequence to encode
-
-    Returns:
-        A list of tuples that represent the encoding for the sequence
-    """
-    seq = [(k, sum(1 for _ in g)) for k, g in groupby(seq)]
-    return [(c, rep) if rep > 1 else (c,) for c, rep in seq]