+++ tests/integration/test_provider.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import pytest
 
-from tergite.services.configs import ProviderAccount
+from tergite.services.configs import AccountInfo
 from tergite.types import provider_factory as factory
 from tests.utils.env import is_end_to_end
 from tests.utils.fixtures import load_json_fixture
@@ -15,10 +15,10 @@ _PROVIDER_ACCOUNTS = load_json_fixture("provider_accounts.json")
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_use_provider_account(account_data, mock_tergiterc):
     """use_provider_account without save option returns the given provider account"""
-    account = ProviderAccount(**account_data)
+    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
     provider = _tergite.use_provider_account(account)
-    assert provider.provider_account == account
+    assert provider.account == account
 
 
 @pytest.mark.skipif(is_end_to_end(), reason="is not end-to-end test")
@@ -35,21 +35,21 @@ def test_use_provider_account_save(account_data, mock_tergiterc):
     for k, v in extras.items():
         expected_config += f"{k} = {v}\n"
 
-    account = ProviderAccount(**account_data)
+    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
     provider = _tergite.use_provider_account(account, save=True)
     with open(mock_tergiterc, "r") as file:
         actual_config = file.read()
 
     assert expected_config in actual_config
-    assert provider.provider_account == account
+    assert provider.account == account
 
 
 @pytest.mark.skipif(is_end_to_end(), reason="is not end-to-end test")
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_get_provider(account_data, mock_tergiterc):
     """get_provider returns the provider with the given service name"""
-    account = ProviderAccount(**account_data)
+    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
     expected = _tergite.use_provider_account(account, save=True)
     got = _tergite.get_provider(account.service_name)
@@ -60,7 +60,7 @@ def test_get_provider(account_data, mock_tergiterc):
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_get_provider_non_existing(account_data, mock_tergiterc):
     """get_provider for non-existing service returns first provider"""
-    account = ProviderAccount(**account_data)
+    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
     _tergite.use_provider_account(account, save=True)
     provider = _tergite.get_provider(datetime.now().isoformat())
@@ -69,4 +69,4 @@ def test_get_provider_non_existing(account_data, mock_tergiterc):
         conf = file.read()
 
     first_service_name = re.search(r"\[service\s+(.+?)]\n", conf).group(1)
-    assert provider.provider_account.service_name == first_service_name
+    assert provider.account.service_name == first_service_name